% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\docType{class}
\name{DataSet-class}
\alias{DataSet-class}
\title{This class stores the information about a dataset.}
\description{
This class stores the information about a dataset.
}
\section{Slots}{

\describe{
\item{\code{path}}{Path to the dataset, typically of the form
\code{"/fastgenomics/data/dataset_xxxx"}}

\item{\code{id}}{Id of the dataset.  Defined as the \code{xxxx} part of the directory
\code{".../dataset_xxxx"} under which the dataset is stored.)}

\item{\code{metadata}}{Information extracted from \code{.../dataset_xxxx/dataset_info.json}}

\item{\code{file}}{Absolute path to the file under which the dataset is stored, something
like \code{".../dataset_xxxx/data.loom"}.}
}}

