% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{DataSet}
\alias{DataSet}
\title{constructor for the DataSet Object, takes dataset path and extracts all information
from it.}
\usage{
DataSet(path)
}
\description{
constructor for the DataSet Object, takes dataset path and extracts all information
from it.
}
