% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ds_info}
\alias{ds_info}
\title{Get information on all available datasets in this analysis.}
\usage{
ds_info(
  ds = NULL,
  pretty = NULL,
  output = NULL,
  data_dir = DATA_DIR,
  ignore_empty = T
)
}
\arguments{
\item{ds}{A single dataset ID or dataset title. If set, only this dataset will be displayed.}

\item{pretty}{Boolean whether to display some nicely formatted output, by default True}

\item{output}{Boolean whether to return a DataFrame or not, by default True}

\item{data_dir}{The directory containing sub-folders of the \code{"dataset_xxxx"}
form.  Defaults to the directory provided by the FASTGenomics platform.}

\item{ignore_empty}{Whether to ignore if no datasets are attached or not.
Defaults to TRUE.}
}
\value{
A data frame containing all, or a single dataset (depends on \code{ds} and \code{output})
}
\description{
The optional argument available under \code{data_dir}.  The function then looks for
all directories matching the \code{".*/dataset_\\d{4}$"} pattern and constructs a
data frame with all data sets.
}
\examples{
\donttest{
dsets <- ds_info()
dsets <- ds_info('Test loom data')
}

}
