% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{expr_file_infos_to_list}
\alias{expr_file_infos_to_list}
\title{Returns DataFileInfos as a vecotor or as a string.}
\usage{
expr_file_infos_to_list(single_ds_df, key, as_string = F)
}
\arguments{
\item{single_ds_df}{A dataframe with a single dataset.}

\item{key}{The column to convert (e.g. "expressionDataFileInfos").}

\item{as_string}{Whether to return a formatted string.}
}
\value{
A vecotr or a formatted string.
}
\description{
Returns DataFileInfos as a vecotor or as a string.
}
