% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_datasets}
\alias{get_datasets}
\title{Lists datasets provided by FASTGenomics.}
\usage{
get_datasets(data_dir = DATA_DIR)
}
\arguments{
\item{data_dir}{The directory containing sub-folders of the \code{"dataset_xxxx"}
form.  Defaults to the directory provided by the FASTGenomics platform.}
}
\value{
List of datasets (as \code{DataSet} objects) indexed by dataset id.
}
\description{
The optional argument available under \code{data_dir}.  The function then looks for
all directories matching the \code{".*/dataset_\\d{4}$"} pattern and constructs a
\code{DataSet} object from each of them.
}
\examples{
\donttest{
dsets_list <- get_datasets()
dsets_list[[1]]  # gives you the dataset with id = 1
}

}
