% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_datasets_df}
\alias{get_datasets_df}
\title{Construct a dataframe of alla vailabe datasets}
\usage{
get_datasets_df(data_dir = DATA_DIR, ignore_empty = T)
}
\arguments{
\item{data_dir}{The directory containing sub-folders of the \code{"dataset_xxxx"}
form.  Defaults to the directory provided by the FASTGenomics platform.}

\item{ignore_empty}{Whether to ignore if no datasets are attached or not.
Defaults to TRUE.}
}
\value{
A dataframe containing all dataset metadata
}
\description{
Construct a dataframe of alla vailabe datasets
}
