% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_ds_paths}
\alias{get_ds_paths}
\title{Get paths of all available datasets}
\usage{
get_ds_paths(data_dir, ignore_empty = T)
}
\arguments{
\item{data_dir}{The directory containing sub-folders of the \code{"dataset_xxxx"}
form. Defaults to the directory provided by the FASTGenomics platform.}

\item{ignore_empty}{Whether to ignore if no datasets are attached or not.
Defaults to TRUE.}
}
\value{
A list of dataset directories
}
\description{
Get paths of all available datasets
}
