% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{list_datasets}
\alias{list_datasets}
\title{Lists data sets provided by FASTGenomics.  The optional argument available under
\code{data_dir}.  The function then looks for all directories matching the
\code{".*/dataset_\\d{4}$"} pattern and constructs a \code{DataSet} object from each
of them.}
\usage{
list_datasets(data_dir = DATA_DIR)
}
\arguments{
\item{data_dir}{The directory containing sub-folders of the \code{"dataset_xxxx"}
form.  Defaults to the directory provided by the FASTGenomics platform.}
}
\value{
List of data sets (as \code{DataSet} objects) indexed by data set id.
}
\description{
Lists data sets provided by FASTGenomics.  The optional argument available under
\code{data_dir}.  The function then looks for all directories matching the
\code{".*/dataset_\\d{4}$"} pattern and constructs a \code{DataSet} object from each
of them.
}
\examples{
dsets_list <- list_datasets()
dsets_list[[1]]  # gives you the data set with id = 1

}
