% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{load_data}
\alias{load_data}
\title{Loads a single dataset to a Seurat object.}
\usage{
load_data(
  ds,
  data_dir = DATA_DIR,
  additional_readers = list(),
  experimental_readers = F,
  as_format
)
}
\arguments{
\item{ds}{A single dataset ID or dataset title.}

\item{data_dir}{The directory containing sub-folders of the \code{"dataset_xxxx"}
form.  Defaults to the directory provided by the FASTGenomics platform.}

\item{additional_readers}{List to specify your own readers for the specific dataset
format. Still experimental, by default an empty list}

\item{experimental_readers}{Boolean whether FASTGenomics in-house experimental
readers should be used. By default set to False.}

\item{as_format}{Specifies which reader should be uses for this dataset. Overwrites the
auto-detection of the format. Possible parameters are the file extensions of our
supported data formats: "h5ad", "h5", "hdf5", "loom", "rds", "csv", "tsv".}
}
\value{
dataset loaded as a Seurat Object
}
\description{
If there are multiple datasets available you need to specify one by setting
\code{ds} to a dataset \code{id} or dataset \code{title}.
To get an overview of availabe dataset use the fucntion \code{ds_info}.
}
\examples{
\donttest{
load_data()                    # returns the Seurat object if only one dataset is available
load_data('Test loom data')    # returns the Seurat object from dataset specified by title
}

}
