% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers_old.R
\name{matrix_to_seurat_old}
\alias{matrix_to_seurat_old}
\title{a helper function that constructs a seurat object from an expression matrix, cell &
gene metadata.  Used to read loom and AnnData files}
\usage{
matrix_to_seurat_old(matrix, cell_metadata, gene_metadata)
}
\description{
a helper function that constructs a seurat object from an expression matrix, cell &
gene metadata.  Used to read loom and AnnData files
}
