% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers_old.R
\name{read_10xhdf5_to_seurat_old}
\alias{read_10xhdf5_to_seurat_old}
\title{Read 10x hdf5 dataset into seurat.}
\usage{
read_10xhdf5_to_seurat_old(data_set)
}
\description{
Read 10x hdf5 dataset into seurat.
}
