% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_10xmtx_to_seurat}
\alias{read_10xmtx_to_seurat}
\title{Read 10x mtx (mex) dataset to seurat.}
\usage{
read_10xmtx_to_seurat(ds_file)
}
\arguments{
\item{ds_file}{The path to the dataset file.}
}
\value{
A seurat Object.
}
\description{
Read 10x mtx (mex) dataset to seurat.
}
