% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers_old.R
\name{read_10xmtx_to_seurat_old}
\alias{read_10xmtx_to_seurat_old}
\title{Read 10x mtx (mex) dataset to seurat.}
\usage{
read_10xmtx_to_seurat_old(data_set)
}
\description{
Read 10x mtx (mex) dataset to seurat.
}
