% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_anndata_to_seurat}
\alias{read_anndata_to_seurat}
\title{Read AnnData to Seurat.
Importing AnnData is not generally available in Seurat v3
mport of AnnData only works if there is a CSR matrix in the AnnData object.
For your convenience the FASTGenomics team provides this beta loading routine.}
\usage{
read_anndata_to_seurat(data_set)
}
\description{
Read AnnData to Seurat.
Importing AnnData is not generally available in Seurat v3
mport of AnnData only works if there is a CSR matrix in the AnnData object.
For your convenience the FASTGenomics team provides this beta loading routine.
}
