% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_anndata_to_seurat_exp}
\alias{read_anndata_to_seurat_exp}
\title{Read AnnData to Seurat with experimental FASTGenomics reader.
Importing AnnData is not always working as expected in Seurat v3
For your convenience the FASTGenomics team provides this beta loading routine.
This custom reader, however, only reads the \code{.X}, \code{.obs} and \code{.var} components of the AnnData object.}
\usage{
read_anndata_to_seurat_exp(ds_file)
}
\description{
Read AnnData to Seurat with experimental FASTGenomics reader.
Importing AnnData is not always working as expected in Seurat v3
For your convenience the FASTGenomics team provides this beta loading routine.
This custom reader, however, only reads the \code{.X}, \code{.obs} and \code{.var} components of the AnnData object.
}
