% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_dataset}
\alias{read_dataset}
\title{Reads a single data set from a given path}
\usage{
read_dataset(path, readers = DEFAULT_READERS)
}
\arguments{
\item{path}{Location of the data set}
}
\description{
Reads a single data set from a given path
}
