% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_densecsv_to_seurat}
\alias{read_densecsv_to_seurat}
\title{Read dense matrix in csv form}
\usage{
read_densecsv_to_seurat(ds_file)
}
\arguments{
\item{ds_file}{The path to the dataset file.}
}
\value{
A seurat Object.
}
\description{
Read dense matrix in csv form
}
