% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_densemat_to_seurat}
\alias{read_densemat_to_seurat}
\title{Read dense matrix
here we need to unpack the dataset before reading it}
\usage{
read_densemat_to_seurat(ds_file, sep)
}
\arguments{
\item{ds_file}{The path to the dataset file.}

\item{sep}{The column seperator (e.g. tab or comma)}
}
\value{
A seurat Object.
}
\description{
Read dense matrix
here we need to unpack the dataset before reading it
}
