% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers_old.R
\name{read_densemat_to_seurat_old}
\alias{read_densemat_to_seurat_old}
\title{Read dense matrix
here we need to unpack the dataset before reading it}
\usage{
read_densemat_to_seurat_old(data_set, sep)
}
\description{
Read dense matrix
here we need to unpack the dataset before reading it
}
