% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_loom}
\alias{read_loom}
\title{Unfortunaly there is no Seurat function that straight up reads the loom file into a
Seurat object so we have to implement our own.  This one only reads the count table
and the row/col attributes that can fit in a data frame structure (i.e.. all higher
dimensional attributes are discarded).  To keep the implementation simple we read the
whole object into memory, including the dense count matrix.  This could be a
potential bottleneck for larger data sets but can be optimized later.}
\usage{
read_loom(directory, manifest)
}
\description{
Unfortunaly there is no Seurat function that straight up reads the loom file into a
Seurat object so we have to implement our own.  This one only reads the count table
and the row/col attributes that can fit in a data frame structure (i.e.. all higher
dimensional attributes are discarded).  To keep the implementation simple we read the
whole object into memory, including the dense count matrix.  This could be a
potential bottleneck for larger data sets but can be optimized later.
}
