% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers_old.R
\name{read_loom_to_seurat_exp_old}
\alias{read_loom_to_seurat_exp_old}
\title{Read a loom dataset with the experimental FASTGenomics reader.
Importing loom files is currently unavailable in Seurat v3.
This beta loader provided by FASTGenomics only reads the count table
and the row/col attributes that can fit in a data frame structure (i.e. all higher
dimensional attributes are discarded). To keep the implementation simple we read the
whole object into memory, including the dense count matrix.  This could be a
potential bottleneck for larger datasets.}
\usage{
read_loom_to_seurat_exp_old(data_set)
}
\description{
Read a loom dataset with the experimental FASTGenomics reader.
Importing loom files is currently unavailable in Seurat v3.
This beta loader provided by FASTGenomics only reads the count table
and the row/col attributes that can fit in a data frame structure (i.e. all higher
dimensional attributes are discarded). To keep the implementation simple we read the
whole object into memory, including the dense count matrix.  This could be a
potential bottleneck for larger datasets.
}
