% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers_old.R
\name{read_loom_to_seurat_old}
\alias{read_loom_to_seurat_old}
\title{Reading a loom dataset currently not implemented in Seurat v3.
For your convenience we implemented experimental readers that you can use by setting "experimental_readers=TRUE"
in \code{\link{read_datasets}} or \code{\link{read_dataset}}.}
\usage{
read_loom_to_seurat_old(data_set)
}
\description{
Reading a loom dataset currently not implemented in Seurat v3.
For your convenience we implemented experimental readers that you can use by setting "experimental_readers=TRUE"
in \code{\link{read_datasets}} or \code{\link{read_dataset}}.
}
