% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{select_ds_id}
\alias{select_ds_id}
\title{Select a single dataset from a pandas DataFrame by its ID or title}
\usage{
select_ds_id(ds, df)
}
\arguments{
\item{ds}{A single dataset ID or dataset title for selection}

\item{df}{A data frame from which a single entry is selected, by default None}
}
\value{
A data frame with only the selected dataset
}
\description{
Select a single dataset from a pandas DataFrame by its ID or title
}
