% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{add_metadata}
\alias{add_metadata}
\title{Adds some data from the metadata directly to the meta.data of the seurat object.}
\usage{
add_metadata(seurat, data_set)
}
\description{
Adds some data from the metadata directly to the meta.data of the seurat object.
}
