% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_anndata_to_seurat}
\alias{read_anndata_to_seurat}
\title{Read AnnData to Seurat with Seurat's function \code{\link{ReadH5AD}}.
Please note that this might not work as expected.
For your convenience we implemented importing with SeuratDisk for h5ad files from Anndata>=0.7.3.
Alternatively, you can use our experimental reader by setting "experimental_readers=TRUE"
in \code{\link{load_data}}.}
\usage{
read_anndata_to_seurat(ds_file)
}
\arguments{
\item{ds_file}{The path to the dataset file.}
}
\value{
A seurat Object.
}
\description{
Read AnnData to Seurat with Seurat's function \code{\link{ReadH5AD}}.
Please note that this might not work as expected.
For your convenience we implemented importing with SeuratDisk for h5ad files from Anndata>=0.7.3.
Alternatively, you can use our experimental reader by setting "experimental_readers=TRUE"
in \code{\link{load_data}}.
}
