% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_anndata_to_seurat}
\alias{read_anndata_to_seurat}
\title{this only works if there's a CSR matrix in the AnnData object.  We would be happy to
use the Seurats ReadH5AD function but it's broken.}
\usage{
read_anndata_to_seurat(data_set)
}
\description{
this only works if there's a CSR matrix in the AnnData object.  We would be happy to
use the Seurats ReadH5AD function but it's broken.
}
