% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_anndata_to_seurat}
\alias{read_anndata_to_seurat}
\title{Read AnnData to Seurat with Seurat's function \code{\link{ReadH5AD}}.
Please note that this might not work as expected.
For your convenience we implemented experimental readers that you can use by setting "experimental_readers=TRUE"
in \code{\link{read_datasets}} or \code{\link{read_dataset}}.}
\usage{
read_anndata_to_seurat(data_set)
}
\description{
Read AnnData to Seurat with Seurat's function \code{\link{ReadH5AD}}.
Please note that this might not work as expected.
For your convenience we implemented experimental readers that you can use by setting "experimental_readers=TRUE"
in \code{\link{read_datasets}} or \code{\link{read_dataset}}.
}
