% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_anndata_to_seurat_exp}
\alias{read_anndata_to_seurat_exp}
\title{Read AnnData to Seurat.
Importing AnnData is not always working as expected in Seurat v3
For your convenience the FASTGenomics team provides this beta loading routine.
Import of AnnData only works if there is a CSR matrix in the AnnData object.}
\usage{
read_anndata_to_seurat_exp(data_set)
}
\description{
Read AnnData to Seurat.
Importing AnnData is not always working as expected in Seurat v3
For your convenience the FASTGenomics team provides this beta loading routine.
Import of AnnData only works if there is a CSR matrix in the AnnData object.
}
