% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_dataset}
\alias{read_dataset}
\title{Reads a single dataset.}
\usage{
read_dataset(data_set, additional_readers = list(), experimental_readers = F)
}
\arguments{
\item{data_set}{The dataset to load (passed as a DataSet object)}
}
\value{
dataset loaded as a Seurat Object
}
\description{
Takes a \code{\link{DataSet}} object as an argument. This function should be used
with conjunction to \code{\link{get_datasets}}.
}
\examples{
\donttest{
dsets_list <- get_datasets()
read_dataset(dsets_list[[1]])  # returns the Seurat object constructed from the first dataset
}

}
