% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_datasets}
\alias{read_datasets}
\title{Loads all datasets available for this analysis.}
\usage{
read_datasets(data_sets = get_datasets(data_dir),
  readers = DEFAULT_READERS)
}
\arguments{
\item{data_sets}{Optional argument with a list of the datasets to load.  Typically a
list returned by \code{\link{get_datasets}} filtered to include only specific
datasets.}
}
\value{
A list of datasets loaded into seurat objects, indexed by dataset id (same
indices as in the \code{\link{get_datasets}}).
}
\description{
Optionally, you can provide a list of datasets to load to load only selected ones.
}
\examples{
# loads all datasets as seurat objects
seurat <- read_datasets()

# loads only the first and second datasets
dsets_list <- get_datasets()
seurat <- read_datasets(dsets_list[c(1,2)])

}
