% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_densemat_to_seurat}
\alias{read_densemat_to_seurat}
\title{Read dense matrix
here we need to unpack the dataset before reading it}
\usage{
read_densemat_to_seurat(data_set, sep)
}
\description{
Read dense matrix
here we need to unpack the dataset before reading it
}
