% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_loom_to_seurat}
\alias{read_loom_to_seurat}
\title{Reading a loom dataset currently not implemented in Seurat v3.
For your convenience we implemented experimental readers that you can use by setting "experimental_readers=TRUE"
in \code{\link{load_data}}.}
\usage{
read_loom_to_seurat(ds_file)
}
\description{
Reading a loom dataset currently not implemented in Seurat v3.
For your convenience we implemented experimental readers that you can use by setting "experimental_readers=TRUE"
in \code{\link{load_data}}.
}
