% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_loom_to_seurat_exp}
\alias{read_loom_to_seurat_exp}
\title{Importing loom files is currently unavailable in Seurat v3
This beta loader provided by our team only reads the count table
and the row/col attributes that can fit in a data frame structure (i.e. all higher
dimensional attributes are discarded). To keep the implementation simple we read the
whole object into memory, including the dense count matrix.  This could be a
potential bottleneck for larger datasets.}
\usage{
read_loom_to_seurat_exp(data_set)
}
\description{
Importing loom files is currently unavailable in Seurat v3
This beta loader provided by our team only reads the count table
and the row/col attributes that can fit in a data frame structure (i.e. all higher
dimensional attributes are discarded). To keep the implementation simple we read the
whole object into memory, including the dense count matrix.  This could be a
potential bottleneck for larger datasets.
}
